Add new SplitView storybook doc
